# =lib Change `Future#submitWithValue` to be tailrec.