@viktorklang 2018-11-15
@hepin1989 Thanks for the PR! What's the impact on performance?