@hepin1989 2018-11-15
I only test it with 10 and 10..
before:
![image](https://user-images.githubusercontent.com/501740/48564428-572c7200-e931-11e8-8890-5ec95d6cb683.png)

after:
![image](https://user-images.githubusercontent.com/501740/48564414-4ed43700-e931-11e8-94e1-746c1b4f09c4.png)