@viktorklang 2018-11-15
@hepin1989 Could you run all of the FutureBenchmark suite with something like 25 warmup and 20 iterations and post the results? (https://github.com/scala/scala/blob/2.13.x/test/benchmarks/src/main/scala/scala/concurrent/FutureBenchmark.scala)