@viktorklang 2018-11-15
You also will want to test with different numbers of threads.