@hepin1989 2018-11-16
after `Thread 1`
```
[info] Benchmark                             (pool)  (recursion)  (threads)   Mode  Cnt       Score     Error   Units
[info] AndThenFutureBenchmark.post              fjp         1024          1  thrpt   30      12.857 ±   0.056  ops/ms
[info] AndThenFutureBenchmark.post              fix         1024          1  thrpt   30      13.356 ±   0.046  ops/ms
[info] AndThenFutureBenchmark.post              fie         1024          1  thrpt   30      17.572 ±   0.074  ops/ms
[info] AndThenFutureBenchmark.post              gbl         1024          1  thrpt   30       2.780 ±   0.116  ops/ms
[info] AndThenFutureBenchmark.pre               fjp         1024          1  thrpt   30      20.096 ±   0.137  ops/ms
[info] AndThenFutureBenchmark.pre               fix         1024          1  thrpt   30      23.207 ±   0.183  ops/ms
[info] AndThenFutureBenchmark.pre               fie         1024          1  thrpt   30       7.777 ±   0.040  ops/ms
[info] AndThenFutureBenchmark.pre               gbl         1024          1  thrpt   30       2.972 ±   0.086  ops/ms
[info] CallbackFutureBenchmark.post             fjp         1024          1  thrpt   30       4.369 ±   0.030  ops/ms
[info] CallbackFutureBenchmark.post             fix         1024          1  thrpt   30       4.348 ±   0.020  ops/ms
[info] CallbackFutureBenchmark.post             fie         1024          1  thrpt   30       7.431 ±   0.029  ops/ms
[info] CallbackFutureBenchmark.post             gbl         1024          1  thrpt   30       4.804 ±   0.048  ops/ms
[info] CallbackFutureBenchmark.pre              fjp         1024          1  thrpt   30       5.299 ±   0.031  ops/ms
[info] CallbackFutureBenchmark.pre              fix         1024          1  thrpt   30       3.384 ±   0.021  ops/ms
[info] CallbackFutureBenchmark.pre              fie         1024          1  thrpt   30       9.314 ±   0.055  ops/ms
[info] CallbackFutureBenchmark.pre              gbl         1024          1  thrpt   30       4.724 ±   0.410  ops/ms
[info] CompleteFutureBenchmark.failure          fjp         1024          1  thrpt   30    1085.124 ±   3.747  ops/ms
[info] CompleteFutureBenchmark.failure          fix         1024          1  thrpt   30    1086.539 ±   2.936  ops/ms
[info] CompleteFutureBenchmark.failure          fie         1024          1  thrpt   30    1037.652 ±  30.109  ops/ms
[info] CompleteFutureBenchmark.failure          gbl         1024          1  thrpt   30    1061.820 ±  13.500  ops/ms
[info] CompleteFutureBenchmark.success          fjp         1024          1  thrpt   30    1111.108 ±   3.152  ops/ms
[info] CompleteFutureBenchmark.success          fix         1024          1  thrpt   30    1110.440 ±   3.399  ops/ms
[info] CompleteFutureBenchmark.success          fie         1024          1  thrpt   30    1111.133 ±   3.158  ops/ms
[info] CompleteFutureBenchmark.success          gbl         1024          1  thrpt   30    1105.021 ±   7.951  ops/ms
[info] CompleteWithFutureBenchmark.failure      fjp         1024          1  thrpt   30     827.939 ±   3.589  ops/ms
[info] CompleteWithFutureBenchmark.failure      fix         1024          1  thrpt   30     823.736 ±   7.750  ops/ms
[info] CompleteWithFutureBenchmark.failure      fie         1024          1  thrpt   30     819.128 ±  15.933  ops/ms
[info] CompleteWithFutureBenchmark.failure      gbl         1024          1  thrpt   30     802.003 ±   5.002  ops/ms
[info] CompleteWithFutureBenchmark.success      fjp         1024          1  thrpt   30     820.476 ±   2.741  ops/ms
[info] CompleteWithFutureBenchmark.success      fix         1024          1  thrpt   30     825.308 ±   2.198  ops/ms
[info] CompleteWithFutureBenchmark.success      fie         1024          1  thrpt   30     820.665 ±   3.049  ops/ms
[info] CompleteWithFutureBenchmark.success      gbl         1024          1  thrpt   30     824.200 ±   2.737  ops/ms
[info] FilterFutureBenchmark.post               fjp         1024          1  thrpt   30      13.796 ±   0.044  ops/ms
[info] FilterFutureBenchmark.post               fix         1024          1  thrpt   30      13.747 ±   0.113  ops/ms
[info] FilterFutureBenchmark.post               fie         1024          1  thrpt   30      21.840 ±   0.120  ops/ms
[info] FilterFutureBenchmark.post               gbl         1024          1  thrpt   30       2.788 ±   0.076  ops/ms
[info] FilterFutureBenchmark.pre                fjp         1024          1  thrpt   30      24.810 ±   0.131  ops/ms
[info] FilterFutureBenchmark.pre                fix         1024          1  thrpt   30      25.449 ±   0.087  ops/ms
[info] FilterFutureBenchmark.pre                fie         1024          1  thrpt   30       8.011 ±   0.153  ops/ms
[info] FilterFutureBenchmark.pre                gbl         1024          1  thrpt   30       2.987 ±   0.123  ops/ms
[info] FirstCompletedOfFutureBenchmark.post     fjp         1024          1  thrpt   30       5.588 ±   0.012  ops/ms
[info] FirstCompletedOfFutureBenchmark.post     fix         1024          1  thrpt   30       4.846 ±   0.024  ops/ms
[info] FirstCompletedOfFutureBenchmark.post     fie         1024          1  thrpt   30       7.532 ±   0.033  ops/ms
[info] FirstCompletedOfFutureBenchmark.post     gbl         1024          1  thrpt   30       5.835 ±   0.075  ops/ms
[info] FirstCompletedOfFutureBenchmark.pre      fjp         1024          1  thrpt   30       5.376 ±   0.042  ops/ms
[info] FirstCompletedOfFutureBenchmark.pre      fix         1024          1  thrpt   30       4.768 ±   0.023  ops/ms
[info] FirstCompletedOfFutureBenchmark.pre      fie         1024          1  thrpt   30       8.280 ±   0.058  ops/ms
[info] FirstCompletedOfFutureBenchmark.pre      gbl         1024          1  thrpt   30       5.924 ±   0.072  ops/ms
[info] FlatMapFutureBenchmark.post              fjp         1024          1  thrpt   30      12.323 ±   0.043  ops/ms
[info] FlatMapFutureBenchmark.post              fix         1024          1  thrpt   30      12.461 ±   0.033  ops/ms
[info] FlatMapFutureBenchmark.post              fie         1024          1  thrpt   30      17.392 ±   0.080  ops/ms
[info] FlatMapFutureBenchmark.post              gbl         1024          1  thrpt   30       2.622 ±   0.027  ops/ms
[info] FlatMapFutureBenchmark.pre               fjp         1024          1  thrpt   30      19.123 ±   0.144  ops/ms
[info] FlatMapFutureBenchmark.pre               fix         1024          1  thrpt   30      16.156 ±   0.264  ops/ms
[info] FlatMapFutureBenchmark.pre               fie         1024          1  thrpt   30       7.693 ±   0.049  ops/ms
[info] FlatMapFutureBenchmark.pre               gbl         1024          1  thrpt   30       2.647 ±   0.083  ops/ms
[info] LoopFutureBenchmark.post                 fjp         1024          1  thrpt   30       5.479 ±   0.040  ops/ms
[info] LoopFutureBenchmark.post                 fix         1024          1  thrpt   30       5.744 ±   0.057  ops/ms
[info] LoopFutureBenchmark.post                 fie         1024          1  thrpt   30       7.254 ±   0.027  ops/ms
[info] LoopFutureBenchmark.post                 gbl         1024          1  thrpt   30       0.773 ±   0.089  ops/ms
[info] LoopFutureBenchmark.pre                  fjp         1024          1  thrpt   30      10.331 ±   0.049  ops/ms
[info] LoopFutureBenchmark.pre                  fix         1024          1  thrpt   30      11.211 ±   0.039  ops/ms
[info] LoopFutureBenchmark.pre                  fie         1024          1  thrpt   30      13.176 ±   0.080  ops/ms
[info] LoopFutureBenchmark.pre                  gbl         1024          1  thrpt   30       2.043 ±   0.070  ops/ms
[info] MapFutureBenchmark.post                  fjp         1024          1  thrpt   30      14.374 ±   0.071  ops/ms
[info] MapFutureBenchmark.post                  fix         1024          1  thrpt   30      14.457 ±   0.128  ops/ms
[info] MapFutureBenchmark.post                  fie         1024          1  thrpt   30      20.449 ±   0.114  ops/ms
[info] MapFutureBenchmark.post                  gbl         1024          1  thrpt   30       3.014 ±   0.066  ops/ms
[info] MapFutureBenchmark.pre                   fjp         1024          1  thrpt   30      21.736 ±   0.046  ops/ms
[info] MapFutureBenchmark.pre                   fix         1024          1  thrpt   30      25.881 ±   0.069  ops/ms
[info] MapFutureBenchmark.pre                   fie         1024          1  thrpt   30       8.765 ±   0.035  ops/ms
[info] MapFutureBenchmark.pre                   gbl         1024          1  thrpt   30       3.028 ±   0.047  ops/ms
[info] NoopFutureBenchmark.post                 fjp         1024          1  thrpt   30   34058.604 ± 167.239  ops/ms
[info] NoopFutureBenchmark.post                 fix         1024          1  thrpt   30   34108.070 ± 145.814  ops/ms
[info] NoopFutureBenchmark.post                 fie         1024          1  thrpt   30   34073.384 ± 164.264  ops/ms
[info] NoopFutureBenchmark.post                 gbl         1024          1  thrpt   30   34078.933 ± 183.273  ops/ms
[info] NoopFutureBenchmark.pre                  fjp         1024          1  thrpt   30  167330.963 ± 782.170  ops/ms
[info] NoopFutureBenchmark.pre                  fix         1024          1  thrpt   30  167922.879 ± 789.868  ops/ms
[info] NoopFutureBenchmark.pre                  fie         1024          1  thrpt   30  168079.888 ± 769.467  ops/ms
[info] NoopFutureBenchmark.pre                  gbl         1024          1  thrpt   30  168068.883 ± 609.567  ops/ms
[info] RecoverFutureBenchmark.post              fjp         1024          1  thrpt   30       8.230 ±   0.028  ops/ms
[info] RecoverFutureBenchmark.post              fix         1024          1  thrpt   30       8.236 ±   0.074  ops/ms
[info] RecoverFutureBenchmark.post              fie         1024          1  thrpt   30       9.377 ±   0.041  ops/ms
[info] RecoverFutureBenchmark.post              gbl         1024          1  thrpt   30       2.062 ±   0.033  ops/ms
[info] RecoverFutureBenchmark.pre               fjp         1024          1  thrpt   30      10.156 ±   0.034  ops/ms
[info] RecoverFutureBenchmark.pre               fix         1024          1  thrpt   30      10.476 ±   0.119  ops/ms
[info] RecoverFutureBenchmark.pre               fie         1024          1  thrpt   30       2.746 ±   0.044  ops/ms
[info] RecoverFutureBenchmark.pre               gbl         1024          1  thrpt   30       2.128 ±   0.033  ops/ms
[info] RecoverWithFutureBenchmark.post          fjp         1024          1  thrpt   30      10.518 ±   0.047  ops/ms
[info] RecoverWithFutureBenchmark.post          fix         1024          1  thrpt   30      10.882 ±   0.036  ops/ms
[info] RecoverWithFutureBenchmark.post          fie         1024          1  thrpt   30      11.838 ±   0.048  ops/ms
[info] RecoverWithFutureBenchmark.post          gbl         1024          1  thrpt   30       2.437 ±   0.075  ops/ms
[info] RecoverWithFutureBenchmark.pre           fjp         1024          1  thrpt   30      15.405 ±   0.121  ops/ms
[info] RecoverWithFutureBenchmark.pre           fix         1024          1  thrpt   30      14.093 ±   0.028  ops/ms
[info] RecoverWithFutureBenchmark.pre           fie         1024          1  thrpt   30       7.041 ±   0.047  ops/ms
[info] RecoverWithFutureBenchmark.pre           gbl         1024          1  thrpt   30       2.570 ±   0.055  ops/ms
[info] SequenceFutureBenchmark.post             fjp         1024          1  thrpt   30       4.087 ±   0.015  ops/ms
[info] SequenceFutureBenchmark.post             fix         1024          1  thrpt   30       4.103 ±   0.017  ops/ms
[info] SequenceFutureBenchmark.post             fie         1024          1  thrpt   30       6.181 ±   0.023  ops/ms
[info] SequenceFutureBenchmark.post             gbl         1024          1  thrpt   30       0.274 ±   0.003  ops/ms
[info] SequenceFutureBenchmark.pre              fjp         1024          1  thrpt   30       4.756 ±   0.011  ops/ms
[info] SequenceFutureBenchmark.pre              fix         1024          1  thrpt   30       4.881 ±   0.060  ops/ms
[info] SequenceFutureBenchmark.pre              fie         1024          1  thrpt   30       3.543 ±   0.030  ops/ms
[info] SequenceFutureBenchmark.pre              gbl         1024          1  thrpt   30       0.284 ±   0.009  ops/ms
[info] TransformFutureBenchmark.post            fjp         1024          1  thrpt   30      12.609 ±   0.064  ops/ms
[info] TransformFutureBenchmark.post            fix         1024          1  thrpt   30      13.100 ±   0.061  ops/ms
[info] TransformFutureBenchmark.post            fie         1024          1  thrpt   30      16.781 ±   0.107  ops/ms
[info] TransformFutureBenchmark.post            gbl         1024          1  thrpt   30       2.970 ±   0.047  ops/ms
[info] TransformFutureBenchmark.pre             fjp         1024          1  thrpt   30      25.865 ±   0.102  ops/ms
[info] TransformFutureBenchmark.pre             fix         1024          1  thrpt   30      27.133 ±   0.076  ops/ms
[info] TransformFutureBenchmark.pre             fie         1024          1  thrpt   30       8.571 ±   0.039  ops/ms
[info] TransformFutureBenchmark.pre             gbl         1024          1  thrpt   30       2.991 ±   0.078  ops/ms
[info] TransformWithFutureBenchmark.post        fjp         1024          1  thrpt   30      13.600 ±   0.060  ops/ms
[info] TransformWithFutureBenchmark.post        fix         1024          1  thrpt   30      13.658 ±   0.065  ops/ms
[info] TransformWithFutureBenchmark.post        fie         1024          1  thrpt   30      18.046 ±   0.066  ops/ms
[info] TransformWithFutureBenchmark.post        gbl         1024          1  thrpt   30       2.666 ±   0.056  ops/ms
[info] TransformWithFutureBenchmark.pre         fjp         1024          1  thrpt   30      23.225 ±   0.064  ops/ms
[info] TransformWithFutureBenchmark.pre         fix         1024          1  thrpt   30      23.506 ±   0.086  ops/ms
[info] TransformWithFutureBenchmark.pre         fie         1024          1  thrpt   30       8.049 ±   0.039  ops/ms
[info] TransformWithFutureBenchmark.pre         gbl         1024          1  thrpt   30       2.661 ±   0.089  ops/ms
[info] VariousFutureBenchmark.post              fjp         1024          1  thrpt   30       1.536 ±   0.013  ops/ms
[info] VariousFutureBenchmark.post              fix         1024          1  thrpt   30       1.461 ±   0.027  ops/ms
[info] VariousFutureBenchmark.post              fie         1024          1  thrpt   30       2.201 ±   0.017  ops/ms
[info] VariousFutureBenchmark.post              gbl         1024          1  thrpt   30       0.167 ±   0.009  ops/ms
[info] VariousFutureBenchmark.pre               fjp         1024          1  thrpt   30       2.213 ±   0.005  ops/ms
[info] VariousFutureBenchmark.pre               fix         1024          1  thrpt   30       2.140 ±   0.103  ops/ms
[info] VariousFutureBenchmark.pre               fie         1024          1  thrpt   30       0.912 ±   0.005  ops/ms
[info] VariousFutureBenchmark.pre               gbl         1024          1  thrpt   30       0.204 ±   0.002  ops/ms
[info] ZipWithFutureBenchmark.post              fjp         1024          1  thrpt   30       4.412 ±   0.010  ops/ms
[info] ZipWithFutureBenchmark.post              fix         1024          1  thrpt   30       3.840 ±   0.016  ops/ms
[info] ZipWithFutureBenchmark.post              fie         1024          1  thrpt   30       6.943 ±   0.052  ops/ms
[info] ZipWithFutureBenchmark.post              gbl         1024          1  thrpt   30       0.275 ±   0.012  ops/ms
[info] ZipWithFutureBenchmark.pre               fjp         1024          1  thrpt   30       5.292 ±   0.015  ops/ms
[info] ZipWithFutureBenchmark.pre               fix         1024          1  thrpt   30       5.586 ±   0.014  ops/ms
[info] ZipWithFutureBenchmark.pre               fie         1024          1  thrpt   30       4.492 ±   0.019  ops/ms
[info] ZipWithFutureBenchmark.pre               gbl         1024          1  thrpt   30       0.285 ±   0.003  ops/ms
```
before Thread 1

```
[info] Benchmark                             (pool)  (recursion)  (threads)   Mode  Cnt       Score      Error   Units
[info] AndThenFutureBenchmark.post              fjp         1024          1  thrpt   30      12.507 ±    0.453  ops/ms
[info] AndThenFutureBenchmark.post              fix         1024          1  thrpt   30      12.085 ±    0.692  ops/ms
[info] AndThenFutureBenchmark.post              fie         1024          1  thrpt   30      17.144 ±    0.818  ops/ms
[info] AndThenFutureBenchmark.post              gbl         1024          1  thrpt   30       2.577 ±    0.185  ops/ms
[info] AndThenFutureBenchmark.pre               fjp         1024          1  thrpt   30      24.231 ±    0.367  ops/ms
[info] AndThenFutureBenchmark.pre               fix         1024          1  thrpt   30      20.617 ±    1.632  ops/ms
[info] AndThenFutureBenchmark.pre               fie         1024          1  thrpt   30       7.262 ±    0.195  ops/ms
[info] AndThenFutureBenchmark.pre               gbl         1024          1  thrpt   30       2.309 ±    0.179  ops/ms
[info] CallbackFutureBenchmark.post             fjp         1024          1  thrpt   30       4.344 ±    0.031  ops/ms
[info] CallbackFutureBenchmark.post             fix         1024          1  thrpt   30       3.990 ±    0.116  ops/ms
[info] CallbackFutureBenchmark.post             fie         1024          1  thrpt   30       7.905 ±    0.095  ops/ms
[info] CallbackFutureBenchmark.post             gbl         1024          1  thrpt   30       4.412 ±    0.268  ops/ms
[info] CallbackFutureBenchmark.pre              fjp         1024          1  thrpt   30       4.437 ±    0.195  ops/ms
[info] CallbackFutureBenchmark.pre              fix         1024          1  thrpt   30       3.942 ±    0.018  ops/ms
[info] CallbackFutureBenchmark.pre              fie         1024          1  thrpt   30       6.902 ±    0.029  ops/ms
[info] CallbackFutureBenchmark.pre              gbl         1024          1  thrpt   30       4.645 ±    0.227  ops/ms
[info] CompleteFutureBenchmark.failure          fjp         1024          1  thrpt   30    1025.434 ±   22.721  ops/ms
[info] CompleteFutureBenchmark.failure          fix         1024          1  thrpt   30    1053.610 ±    4.557  ops/ms
[info] CompleteFutureBenchmark.failure          fie         1024          1  thrpt   30     956.527 ±   55.113  ops/ms
[info] CompleteFutureBenchmark.failure          gbl         1024          1  thrpt   30     981.149 ±   16.791  ops/ms
[info] CompleteFutureBenchmark.success          fjp         1024          1  thrpt   30     972.875 ±   34.278  ops/ms
[info] CompleteFutureBenchmark.success          fix         1024          1  thrpt   30     994.433 ±   48.654  ops/ms
[info] CompleteFutureBenchmark.success          fie         1024          1  thrpt   30    1047.752 ±    8.764  ops/ms
[info] CompleteFutureBenchmark.success          gbl         1024          1  thrpt   30    1008.447 ±   29.140  ops/ms
[info] CompleteWithFutureBenchmark.failure      fjp         1024          1  thrpt   30     733.850 ±   19.503  ops/ms
[info] CompleteWithFutureBenchmark.failure      fix         1024          1  thrpt   30     707.467 ±   20.864  ops/ms
[info] CompleteWithFutureBenchmark.failure      fie         1024          1  thrpt   30     733.525 ±   10.662  ops/ms
[info] CompleteWithFutureBenchmark.failure      gbl         1024          1  thrpt   30     713.597 ±   17.057  ops/ms
[info] CompleteWithFutureBenchmark.success      fjp         1024          1  thrpt   30     735.958 ±   15.038  ops/ms
…